requirements.txt
asgiref==3.11.0
babel==2.18.0
brotli==1.2.0
cffi==2.0.0
cssselect2==0.9.0
Django==6.0.3
django-debug-toolbar==6.2.0
djangorestframework==3.16.1
dotenv==0.9.9
et_xmlfile==2.0.0
fonttools==4.62.1
gunicorn==25.0.3
html5lib==1.1
iniconfig==2.3.0
openpyxl==3.1.5
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
psycopg2-binary==2.9.11
pycparser==3.0
pydyf==0.10.0
Pygments==2.19.2
pyphen==0.17.2
pytest==9.0.2
pytest-django==4.12.0
python-dotenv==1.2.1
six==1.17.0
sqlparse==0.5.5
tinycss2==1.5.1
weasyprint==60.2
webencodings==0.5.1
whitenoise==6.11.0
zopfli==0.4.1
